fix channel provider closed race condition
